requirements.txt
fastapi
langgraph
playwright
openai
whisper
uvicorn
python-dotenv